+++ Detector/Detector/pch.cpp
@@ -1,6 +1,0 @@
-// pch.cpp: source file corresponding to the precompiled header.
-
-#include "pch.h"
-
-// When you are using precompiled headers, this source file is necessary for
-// compilation to succeed.